+++ packages/datagraphic/stories/scales.stories.js
@@ -1,0 +1,9 @@
+import ScalePointStory from "./scales/ScalePoint.svelte";
+
+export default {
+  title: "Scales",
+};
+
+export const ScalePoint = () => ({
+  Component: ScalePointStory,
+});